create useCovidData reducer within new reducers directory
useCovidData will be in charge of fetching data from the Covid-19 REST API
on application initial rendering.

It exposes whether the request is being processed (and thus `loading` will be
set to true), the data from the response (assuming it was executed successfully)
and the error, if one occured in the process.