create Covid 19 world wide dashboard
The dashboard presents data fetched from https://api.covid19api.com.

During the time the data is being feched, an hourglass spinner
indicates that the application is loading.

Initially, it shows global stats of new and total confirmed cases
and new and total deaths.

In addition, a select component allows the user to see data
from a specific country. Once a country is selected, the dashboard
shows the corresponding data.

If a country is selected, a reset button appears.
When clicked, it resets the dashboard to show the initial global data.

Currently, no error handling is performed.
In the future, if the fetch failed, the error will be presented to the user.